samba: fix dependencies
The samba configure scripts detects the presence of fam and avahi
headers in order to decide whether support for fam and/or avahi should
be built into samba. This patch adds the missing dependencies so that
fam and/or avahi are built before samba if they are selected in buildroot.

Signed-off-by: Sven Neumann <s.neumann@raumfeld.com>