raumfeld: new feature for bootloader: reset_env
When the bootloader starts up, it checks for the presence of the env
variable 'reset_env'. if it exists, the environment is restored from the
defaults hard-coded in the loader itself.

this is now used by the update script to force this procedure after the
bootloader env has been overwritten with zeros.

the particular problem here is that we do not want to have more than one
location to store the bootloader defaults. so letting the loader itself
do the job appears to be safest.