gnutls: update to version 2.10.3
There are newer releases available, but this is the last release that
uses libgcrypt. Newer versions prefer nettle over gcrypt. But then there
are users of gnutls (for example libsoup and glib-networking) that
assume that gnutls uses gcrypt as its crypto backend. They won't compile
with a version of gnutls that uses nettle.

Signed-off-by: Sven Neumann <s.neumann@raumfeld.com>