raumfeld: limit DMA mode for the new disks to UDMA4 (66 MB/sec)
With the SATA adapter the cable detection does not any longer
work. This causes UDMA5 mode (100 MB/sec) to be selected for
the new disks.

Because we are afraid that this is too much on the 44 wire cable,
we select UDMA4 here. This will result in an error with the old
disks because the driver limits itself to UDMA2 if it detects
a 44 wire cable.