udev: add patch to allow synchronous firmware requests
udev 182 doesn't any longer process firmware requests from modules
until the request to load the module has been resolved. This leads
to a timeout of 30 seconds for loading the firmware of modules that
request their firmware synchronously from module_init().

This patch is taken from the udev tree in systemd and solves this
issue. The right fix is of course to use a newer kernel that
implement firmware loading (>= 3.7) or to fix the module so that
it loads the firmware asynchronously. However this is not always
feasible.

Signed-off-by: Sven Neumann <s.neumann@raumfeld.com>