udev: fix patch for firmware loading
Fixes RFS-3787 (Firmware loading broken on old ARM hardware)

The patch introduced with commit 3d3a444c60 was not properly
backported and put the new code into the wrong place. This
is now fixed and this version of the patch matches the
original intent and fixes our firmware loading problem.

Signed-off-by: Sven Neumann <neumann@teufel.de>