raumfeld: add more input tests
Change the key_1, key_2, key_3 tests to actually test KEY_1, KEY_2
and KEY_3 and not the function keys KEY_F1, KEY_F2, KEY_F3. We will
need these tests for the station buttons on the new Raumfeld One.

Add a key_f3 test and change all users of key_3 to use this test
instead.

Add tests for the volume buttons on the Raumfeld Cube.

Part of the fix for RFS-3849