dhcp: add patch to use monotonic time
Some time after we have set the time using ntpdate, the
dhclient will drop the IP and ask the DHCP server for a
new lease. This is also reported by other people, see
for example
http://billauer.co.il/blog/2012/10/dhcp-ip-ntpdate-rtc/

Fix this by changing the dhcp code to use monotonic time
instead of calling gettimeofday().

Signed-off-by: Sven Neumann <neumann@teufel.de>