raumfeld: remove leftover temporary files on startup
GIO creates temporary files using the name ".goutputstream-XXXXXX".
Such files may be left behind if the writing process is killed. To
avoid that we are running out of "disk" space, clean up such files
on boot.

Rename the startup script to 03raumfeld on all platform and remove
the unnecessary explicit mounting.

Cleanup for RFS-3832