raumfeld: am335x-pm-firmware.bin: add version with PER domain workaround
We're not yet able to put the peripheral power domain fully to sleep due
to some bugs in cross-dependant drivers of the Linux kernel.

This firware has a workout to not let that domain in suspend at all,
which for now solves the problem but leaves us with ~0.1 Watt of extra
power consumption. The hack will be reverted once the root of the evil
is extinguished.