opus: allow to select fixed-point implementation
The default of choosing the fixed-point implementation only
if the target hardware does not have floating-point unit is
not always the best choice. Modern ARM hardware has floating
point support but most likely the fixed-point implementation
in Opus will still perform better. In particular if your CPU
supports the NEON extension.

This commit adds two options that allow for better fine-tuning
of opus performance for the target platform.

Signed-off-by: Sven Neumann <neumann@teufel.de>