raumfeld: test_jig: add test for jig
This test assumes a test jig board that is connected to the 20-pin
headers on the baseboard. Each relevant I/O pin is electrically
connected to a 2nd one.

The code walks the array of GPIOs and activates each one of them as
output while keeping the others as inputs. All other pins are measured
and the result is compared to the expected state.

Note that some pins have pull-ups on the S800 module, while others
have pull-downs. Hence, the pinctrl matrix configures the internal pull
configuration to the same setup, and the .pull member of struct
test_gpio denotes which value to expect.

One pin is currently disabled, as the PLL chip seems to drive the line,
even though is configured not to.