samba: use the built-in getaddrinfo() replacement
It appears that getaddrinfo() from glibc does not behave as
expected by the Samba developers. In particular it doesn't
like to be called with "0.0.0.0" and AI_PASSIVE.

The configure script checks the behaviour of getaddrinfo()
but this check can not be performed when cross-compiling.
So instead of assuming that getaddrinfo() behaves as
expected, assume that it wouldn't. This causes the builtin
replacement to be used which provides the expected
behaviour.

Fixes RFS-4484

Signed-off-by: Sven Neumann <neumann@teufel.de>