raumfeld: run USB input test if USB sound-card is connected
Change the end-of-line test so that the presence of a USB sound
card is detected on a Raumfeld Element before the tests start.
In presence of a USB sound card we don't start the tests but
the factory-tests binary which will feed the audio input from
the USB sound card to the speaker.

Part of the changes for RFS-4765