raumfeld: tweak update-uboot-armada.sh
Our current approach to determine whether or not to update U-Boot is
somewhat naive, as we always update as long as the currently flashed version
differs from our image file. One major disadvantage is that with old images,
we override the current bootloader with an older one.

Fix this by keeping track of known U-Boot version by their md5sum. Only
update if we known which version of the bootloader we currently have, and
suspect that in case we don't recognize it, it might be because the image we
booted from is older than the bootloader.