raumfeld: do not load WiFi module in imgrootfs for legacy ARM
Loading the module takes ages as it waits for the firmware to be
loaded which does never happen because udev is not available.
As we do not any longer perform any tests here, we don't need
WiFi at all.

Wiped out some other modules as well which don't seem to be needed.

Fixes RFS-4983