openssh: improve checks for ssh key files
Check if ssh-keygen is executable. Generate a ssh key not only
if the key file does not yet exist but also if it has size zero.

Sometimes we end up with zero size key files because the user
pulled the plug while we were writing the key file. With this
change the key will be regenerated on next boot.