connman: tethering (bridged-ap): added support for a second ethernet interface (eth1).
On Raumfeld Expand 1 it is necessary to remove/add the interface eth1 to/from the tether bridge.
With this change the client devices connected to eth1 receive an IP address via DHCP and are connected to the local network.
The change fixes the problem that sometimes eth0 was missing after enabling the ethernet technology on Raumfeld Expand 1, too.

RFS-5870