gupnp: delay the creation of the GUPnPContext
It appears that connman signals an interface to be 'ready' as
soon as it has an IP. However the routes may not have been
configured correctly at this point. This should actually be
fixed in connman, but for now it seems easier to work around
the problem here.

Workaround for RFS-5884