raumfeld: reduce WiFi transmit power from udev rule
The use of an init script has a potential race condition as we
can not be sure that the WiFi module has been loaded when the
init script is executed (see RFS-5832 for a similar problem).

Move the code to a udev rule that is executed when the wlan0
device is created.

Fixup for RFS-5568