Makefile: staging symlink uses a relative path
A step forward toward a relocatable SDK.

Signed-off-by: Samuel Martin <s.martin49@gmail.com>

Taken from <https://patchwork.ozlabs.org/patch/494784/> by Sam
Thursfield. This might be a bit of a hack... commit
5ad139ebc9916413d25cfe1603da9 suggests that the staging/ symlink is a
deprecated compatibility feature.

However, when using the host/ dir as an SDK, it's necessary to be able
to get a path to the sysroot files, and since the real path name
contains an architecture-specific string it's not simple to calculate
the real path name.