factory-tests: don't start ioboard-audio-pins test
Daniel Konrad told us to remove this test from the ioboard iopins tests because the factories in china report 10% garbage because this test fails often at the moment.
Due to the fact that the audio-line-in test (which is started afterwards) only works if the audio pins are working we don't really need the ioboard-audio-pins test.