Add toplevel CMake file
This wraps Buildroot and the Raumfeld image generation modules,
replacing the build.sh and build-finish.sh shell scripts that were
previously used.

This is a *huge* amount more lines of code... the main benefit is
that the dependencies between the different buildroot modules are
taken into account, so far. This makes it easier to split things up
further in future, and to implement some sort of artifact caching.

Version number is set by passing -DRAUMFELD_VERSION=xx to CMake at
configure-time.