updatecreate.sh: Create all updates for one target type as a single .tar
This is much easier for build logic to deal with, and also needed so we
can store the results in the Artifactory artifact cache easily.

This change also moves the hardware ID information out of
updatecreate.sh and into the caller, and adds a --output-file parameter
to specify the name of the output .tar file.

The downside of this is that the updates need to be unpacked onto a web
server for them to actually be usable by the Raumfeld devices. This will
need to be handled outside the build process from now on.