raumfeld.mk: add cmake parameters that empty cflags injected by cmake.
* Don't add default cmake c/cxxflags after the flags given by
  buildroot, the ones injected by buildroot are more appropriate.
* Set cmake parameter IN_BUILDROOT_CONTEXT=1, such that the core build
  system can detect invocation by buildroot.

This brings the build into a state where the resulting image seems to
work (i.e. all expected raumfeld processes are running).
Cmake injects the linkflag "-rdynamic", which causes the size of /raumfeld
to increase from 17.9M to 19.6M. However, the use of -rdynamic might be
beneficial when creating stack traces, so I left it in for the time being.