Rationalise how debug stripping is done in the toolchain
The toolchain build produces certain files that end up in the target
filesystems: libstdc++, libgcc_s, libc.so.6 and other small bits of GCC
and GLIBC.

The BR2_STRIP_EXCLUDE_FILES setting we had was copied over from the
audioadapter, base and remote-control configs, and was referencing files
that don't actually exist in the toolchain build.

We continue to strip debug info for 'arm' and 'geode' targets where
flash space is precious, but enable debug info for 'armada' because it's
very useful having debug symbols from the libc and libstdc++ when
looking at crash reports. The debug info for libc and libstdc++ takes up
roughly 5MB at present for 'armada'.