Avoid multiple builds of the same Buildroot targets
When CMake's generated Makefiles are run with parallel Make, you
discover https://cmake.org/Bug/view.php?id=10082 and
https://cmake.org/Bug/view.php?id=12311.

To avoid this problem, the images and updates targets need to depend on
the Buildroot *targets*, not the output files. Except that file
dependencies aren't propagated from custom targets, so we *also* need to
depend on the files.

That fixes this sort of disaster where two 'make' instances are run at
once in the same Buildroot directory:

    [ 13%] Building Buildroot config raumfeld/br2-imgrootfs-armada.config to produce /home/sam/teufel/buildroot/build4/buildroot-imgrootfs-armada/images/rootfs.ext2
    [ 16%] Building Buildroot config raumfeld/br2-imgrootfs-armada.config to produce /home/sam/teufel/buildroot/build4/buildroot-imgrootfs-armada/images/rootfs.ext2