Make raumfeld-extract come from its own repo, rather than core.git.
This means that the buildroot-initramfs-* artifacts no longer contain
anything from core.git. If they did, we would need to rebuilt them on
each commit to core.git, which would be wasteful, or we would need to
figure out a way to build raumfeld-extract outside of buildroot and
insert it into an already-generated uImage initramfs, which would be
complicated.