raumfeld: all builds should be done from the feature branch
As we are doing massive changes to the build system that require
changes in the core and buildroot repositories, it is important
that core is always built from the feature branch.

In a couple of days we should be ready to switch this branch
over to the new build system.